// web/src/api/generated/models/suiteResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kick Sim Studio API
 * OpenAPI spec version: 0.2.0
 */
import type { CaseResult } from './caseResult';
import type { Thresholds } from './thresholds';

export interface SuiteResult {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** @nullable */
  cases: CaseResult[] | null;
  endedAt: string;
  failedCases: number;
  name: string;
  passed: boolean;
  passedCases: number;
  passedPercent: number;
  startedAt: string;
  suiteId: string;
  thresholds: Thresholds;
}