// web/src/api/generated/models/workflowResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kick Sim Studio API
 * OpenAPI spec version: 1.0.0
 */
import type { Delivery } from './delivery';

export interface WorkflowResult {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** @nullable */
  deliveries: Delivery[] | null;
  endedAt: string;
  error?: string;
  kind: string;
  name: string;
  passed: boolean;
  scenarioId: string;
  startedAt: string;
}